+++ src/protocol/mod.rs
@@ -60,8 +60,13 @@ where
         }
     }
 
-    pub fn timeout(mut self, timeout: u32) -> Self {
-        self.message.headers.timelimit = (None, Some(timeout));
+    pub fn time_limit(mut self, time_limit: u32) -> Self {
+        self.message.headers.timelimit.1 = Some(time_limit);
+        self
+    }
+
+    pub fn hard_time_limit(mut self, time_limit: u32) -> Self {
+        self.message.headers.timelimit.0 = Some(time_limit);
         self
     }
 
@@ -173,8 +178,12 @@ where
 
         let mut builder = MessageBuilder::<T>::new(id);
 
-        if let Some(timeout) = task_sig.timeout.take() {
-            builder = builder.timeout(timeout);
+        if let Some(time_limit) = task_sig.time_limit.take() {
+            builder = builder.time_limit(time_limit);
+        }
+
+        if let Some(time_limit) = task_sig.hard_time_limit.take() {
+            builder = builder.hard_time_limit(time_limit);
         }
 
         if let Some(countdown) = task_sig.countdown.take() {