+++ src/task/request.rs
@@ -46,7 +46,7 @@ where
     pub reply_to: Option<String>,
 
     /// The time limit (in seconds) allocated for this task to execute.
-    pub timeout: Option<u32>,
+    pub time_limit: Option<u32>,
 }
 
 impl<T> Request<T>
@@ -54,7 +54,7 @@ where
     T: Task,
 {
     pub fn new(m: Message, p: T::Params) -> Self {
-        let timeout = match m.headers.timelimit {
+        let time_limit = match m.headers.timelimit {
             (Some(soft_timelimit), Some(hard_timelimit)) => {
                 Some(std::cmp::min(soft_timelimit, hard_timelimit))
             }
@@ -74,7 +74,7 @@ where
             expires: m.headers.expires,
             hostname: None,
             reply_to: m.properties.reply_to,
-            timeout,
+            time_limit,
         }
     }